pyproject.toml
[project]
name = "fathom"
version = "0.1.0"
description = "Lightweight Sonarr+Radarr replacement with LLM-powered release parsing"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0.30,<2.1",
    "aiosqlite>=0.20",
    "alembic>=1.14,<2.0",
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "httpx>=0.27,<1.0",
    "openai>=1.0",
    "apscheduler>=3.11,<4.0",
    "jinja2>=3.1",
    "lxml>=5.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]

[project.scripts]
fathom = "fathom.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fathom"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
